Rename docker-client to trainer
