+++ docker-client/openapi-codegen.sh
@@ -4,10 +4,10 @@ set -e
 
 cd "$(dirname "$0")"
 
-rm -rf open_pipe_docker_api_client open-pipe-docker-api-client
+rm -rf open_pipe_internal_api_client open-pipe-internal-api-client
 
 poetry run openapi-python-client generate --path ./openapi.json
 
-rm -rf docker_client/api_client 
-mv open-pipe-docker-api-client/open_pipe_docker_api_client docker_client/api_client
-rm -rf open-pipe-docker-api-client
+rm -rf trainer/api_client 
+mv open-pipe-internal-api-client/open_pipe_internal_api_client trainer/api_client
+rm -rf open-pipe-internal-api-client
