+++ docker-client/trainer/api_client/__init__.py
@@ -1,4 +1,4 @@
-""" A client library for accessing OpenPipe Docker API """
+""" A client library for accessing OpenPipe Internal API """
 from .client import AuthenticatedClient, Client
 
 __all__ = (
