+++ docker-client/trainer/api_client/api/default/get_training_info.py
