+++ docker-client/trainer/api_client/models/get_training_info_response_200_base_model.py
